feat:商品实体信息修改

+++ mall-pms/pms-api/src/main/java/com/youlai/mall/pms/entity/PmsSku.java
@@ -15,7 +15,7 @@ public class PmsSku extends BaseEntity {
     private String specification;
     private String barCode;
     private String pic;
-    private Integer originPrice;
-    private Integer price;
+    private Long originPrice;
+    private Long price;
     private Integer stock;
 }
